SEC/T1-Assinatura-digital_Bruno-Leal/requirements.txt
pycryptodome>=3.10.1

reportlab~=4.2.5
pillow~=11.0.0
pypdf~=5.0.1